Add module on interactive list
